+++ README.md
@@ -2,6 +2,14 @@
 
 notion-zh_CN 是对notion的汉化脚本。2.0版本支持网页端(油猴脚本)+桌面端。
 
+## 功能
+
+1. 网页端 ( 油猴脚本 ) + 桌面端 ( win / mac ) 支持最新版本
+  > 未来官方支持中文版也会跟进（如果官方做得好就可以功成身退了）
+
+2. 支持 中文 / **拼音** 快捷键输入（2.1.0支持）
+  > 支持拼音快捷键是便于英文，中文同时输入的时候不用反复切换输入法来保证快捷键输入。
+
 # 为什么要做这个项目？
 
 官方连机器翻译都不愿意做。
@@ -66,6 +74,6 @@ notion-zh_CN 是对notion的汉化脚本。2.0版本支持网页端(油猴脚本
 
 # 大家可以做什么？
 
-1. **优化汉化语言**。都是机器翻译，看到不通畅的句子欢迎提issue/pr直接改了
+1. **优化汉化语言**。都是机器翻译，看到不通畅的句子欢迎提issue/pr直接改了 （修改 **`json/zh.json`** 文件，了解之前，先找到原有的英文，韩文对照一下再更新翻译。）
 2. [x] <del>为什么windows需要使用韩文才能触发汉化</del>
 3. [x] <del>Mac 上有没有相对简单的办法也有时间国际化的能力</del>
