--- buffer
@@ -1,98 +1,0 @@
-"UpdateSidebarFollowControl.following.caption":"发送所有更新和评论的通知"
-"UpdateSidebarFollowControl.following.label":"取消关注"
-"UpdateSidebarFollowControl.notFollowing.caption":"无通知发送"
-"UpdateSidebarFollowControl.notFollowing.label":"不关注"
-"action.languageMode.julia":"朱莉娅"
-"action.languageMode.protobuf":"Protobuf"
-"activity.accessRequested.messageLabel":"{author}您的留言："
-"activity.collectionEdited.header":"您已经编辑了{numberOfAuthors、plural、other{{authorOrAuthors}{collectionViewTitle}视图。}}"
-"activity.collectionViewDeleted.header":"{numberOfAuthors，plural，other{{authorOrAuthors}{collectionTitle}从{collectionViewTitle}删除了{collectionViewTitle}视图。}"
-"allTimeZones.Asia/Saigon":"亚洲/胡志明市"
-"allTimeZones.Etc/Zulu":"其他/祖鲁"
-"appUpdateListener.downloadAndReinstallDesktopApp.message":"要脱机工作并充分利用最新功能，请下载<textlink>并重新安装{isMacApp、select、true{Mac应用}other{Windows应用}}</textlink>。"
-"appUpdateListener.downloadingDesktopApp.message":"正在更新{isMacApp、select、true{Mac应用程序}other{Windows应用程序}}，以便脱机工作({processPercent})。"
-"blockPropertyValue.filterForNumberFormats.placeholder":"搜索数字格式"
-"bookmarkBlock.editBookmark.linkPlaceholder":"粘贴URL(https://.)"
-"collection.numberFormat.dollar":"美元"
-"collection.numberFormat.newTaiwanDollar":"新新台币"
-"collectionFilterMenuFilter.filterOperators.checkbox.order":"{operatorValue}{operator}"
-"collectionFilterMenuFilter.filterOperators.date.order":"{operatorValue}{operator}"
-"collectionFilterMenuFilter.filterOperators.file.order":"属性{operatorValue}{operator}"
-"collectionFilterMenuFilter.filterOperators.multi_select.order":"属性{operatorValue}{operator}"
-"collectionFilterMenuFilter.filterOperators.number.order":"属性{operatorValue}{operator}"
-"collectionFilterMenuFilter.filterOperators.person.order":"属性{operatorValue}{operator}"
-"collectionFilterMenuFilter.filterOperators.relation.order":"{operatorValue}{operator}"
-"collectionFilterMenuFilter.filterOperators.select.order":"属性{operatorValue}{operator}"
-"collectionFilterMenuFilter.filterOperators.text.order":"属性{operatorValue}{operator}"
-"collectionFilterMenuFilter.property.name":"属性"
-"collectionFilterMenuHelpers.operator.stringHasNoAlphabetPrefix":"(禁用)"
-"collectionSortMenu.mobile.sort":"排序{sortNumber}"
-"comment.embeddedFile.placeholder":"正在上载文件."
-"comment.embeddedFile.uploadInProgressMessage":"正在上载文件."
-"configureRelationModal.relationProperty.createModal.selfRelationPrompt.createNewPropertyOption.caption":"<b>在一个属性上显示并且不同步。在</b>“操作”数据库中创建一个“相关任务”属性时，请使用此选项。"
-"configureRelationModal.relationProperty.createModal.selfRelationPrompt.samePropertyOption.caption":"<b>进一步创建新属性并双向同步。在</b>“操作”数据库中分别创建“父任务”和“子任务”属性时，请使用此选项。"
-"confirmDialogInput.incorrectInputError.message":"输入“{requiredInputValue}”以继续。"
-"createPageMenuItem.title.withPageName":"添加新的“{filterText}”页面"
-"database.boardView.uncategorizedColumnTooltip":"<b>{name}</b>属性为空的项目将移到此处。无法删除此组。"
-"database.collectionGroupValue.dateGroup.week":"{year}年{month}月{startDay}日~{endDay}日的周"
-"database.collectionGroupValue.numberGroup.outOfRange":"超出范围"
-"database.collectionGroupValue.numberGroup.range":"从{start}到{end}"
-"database.collectionGroupValue.textGroup.other":"其他"
-"database.formula.function.dateAdd.description":"加上日期。最后一个参数单位可以是“years”、“quarters”、“months”、“weeks”、“days”、“hours”、“minutes”、“seconds”、“milliseconds”。"
-"database.formula.function.dateBetween.description":"返回两个日期之间的时间。最后一个参数单位可以是“years”、“quarters”、“months”、“weeks”、“days”、“hours”、“minutes”、“seconds”、“milliseconds”。"
-"database.formula.function.dateSubtract.description":"从日期减去。最后一个参数单位可以是“years”、“quarters”、“months”、“weeks”、“days”、“hours”、“minutes”、“seconds”、“milliseconds”。"
-"database.formula.function.id.description":"返回每个条目的唯一字符串ID。"
-"database.groupMenu.dateGroupBy.relative":"相对日期"
-"database.groupMenu.numberGroupRange":"组范围"
-"database.groupMenu.numberGroupSize":"所有组"
-"database.groupMenu.sortType.alphabetical":"按字母顺序排列"
-"database.groupMenu.sortType.ascending":"升序"
-"database.groupMenu.sortType.chronological":"按时间顺序"
-"database.groupMenu.sortType.descending":"降序"
-"database.groupMenu.sortType.manual":"手动"
-"database.groupMenu.sortType.reverseAlphabetical":"字母倒序"
-"database.groupMenu.sortType.reverseChronological":"时间倒序"
-"database.groupMenu.textGroupBy.alphabetical":"按字母顺序排列"
-"database.groupMenu.textGroupBy.exact":"按特定单词"
-"database.relationProperty.createModal.selfRelationPrompt.createNewPropertyOption.caption":"<b>无同步。在</b>“操作”数据库中为“相关任务”属性建模时，请使用此选项。"
-"database.relationProperty.createModal.selfRelationPrompt.samePropertyOption.caption":"<b>双向同步。在</b>“任务”数据库中为“父任务”(父任务)和“子任务”属性建模时，请使用此选项。"
-"database.searchRelation.createNewPageLabel":"在{databaseNameWithIcon}<medium>中创建一个</medium>名为{pageName}的<medium>新页面。</medium>"
-"database.selectRelation.menuItem":"选择页面"
-"developerIntegrationCardDropdown.delete.confirmationError":"输入的统一名称不匹配。"
-"developerIntegrationCardDropdown.delete.confirmationMessage":"此公共集成将被禁用，并从所有添加的工作区中删除。要确认，请输入此集成的名称。"
-"developerIntegrationCardDropdown.delete.confirmationTitle":"是否要删除{integrationName}集成？"
-"discussionInput.insertMention.button.tooltip":"输入您想要声明的用户、页面和日期。"
-"loginErrors.csrf.message":"如果您通过链接登录，请在您请求链接的浏览器中打开该链接。"
-"oauthAuthorizationPage.permissionStep.createSpaceLevelIntegrationCheckbox.label":"{workspaceName}与所有成员集成共享"
-"passwordChangeNotificationEmail.removePassword.message":"您可以通过在登录页面上选择“通过电子邮件继续”继续访问Notion。我们将通过电子邮件向您发送临时密码。"
-"passwordChangeNotificationEmail.unintendedChange.message":"如果您没有提出更改请求，请发送电子邮件至team@makenotion.com。转到您的Notion个人资料设置以更改密码，或使用“忘记密码”链接更改密码。"
-"removeUsersFromSpace.nonexistentSpace.error.message":"不存在的工作空间。"
-"removeUsersFromSpace.removingLastAdmin.error.message":"无法删除最后一个管理员。"
-"securitySAMLSettings.securitySection.disableMovingPages.caption":"此选项通过“移动”或“复制”功能防止其他人将页面移动或复制到另一个工作区。"
-"spaceBasicSettings.PublicHomePageSection.input.placeholder":"选择Web上共享的页面"
-"spaceBasicSettings.PublicHomePageSection.title":"公共主页"
-"spaceBasicSettings.changeDomain.cta.text":"直接输入"
-"spaceBasicSettings.domainSection.workspaceDomainInput.joinWorkspace.caption":"如果有允许的电子邮件域，任何人都可以通过{linkText}链接订阅此工作空间。"
-"spaceBasicSettings.domainSection.workspaceDomainInput.viewPublicPage.caption":"您可以在{linkText}上找到在网络上共享的页面。"
-"stripeHelpers.invalidExpiryYear.error.message":"卡过期了。"
-"stripeHelpers.invalidNumber.error.message":"信用卡号码无效。"
-"text.commandsMenuNotOpen.placeholder":"使用命令时输入“/”"
-"topbar.updatesButton.closeSidebartTooltip":"关闭更新侧边栏"
-"topbar.updatesButton.openSidebartTooltip":"打开更新侧边栏"
-"updateSidebar.resolvedToggle.hideResolved":"隐藏已解决的评论。"
-"updateSidebar.resolvedToggle.showResolved":"显示已解决的评论。"
-"updateSidebar.updatesHeader.label":"更新"
-=======
-"UpdateSidebarFollowControl.following.caption":"接收所有更新和评论通知"
-"UpdateSidebarFollowControl.notFollowing.caption":"不接收通知"
-"blockPermissionsSettings.permissionsForBot.personalBot.label":"只有我能看到的状态"
-"blockPermissionsSettings.permissionsForBot.spaceBot.label":"与工作区共享"
-"botMenuItem.spaceBot.label":"与工作区共享"
-"collectionFilterMenuHelpers.operator.stringEndsWith":"以值结尾的数据"
-"discussionInput.sendComment.button":"评论"
-"newBlock.excalidraw.fuzzySearchKeywords":"Excalidraw Xcalidro"
-"newBlock.replit.fuzzySearchKeywords":"宣传册"
-"newBlock.sketch.fuzzySearchKeywords":"草绘Sketch"
-"subscriptionErrors.couponAlreadyApplied":"这是已经使用的优惠券。"
-"updateSidebar.resolvedToggle.hideResolvedTooltip":"隐藏已解决的评论"
-"updateSidebar.resolvedToggle.showResolvedTooltip":"查看已解决的评论"
