+++ package.json
@@ -12,5 +12,10 @@
     "jsdom": "^17.0.0",
     "prettier": "^2.3.2",
     "tencentcloud-sdk-nodejs": "^4.0.183"
-  }
+  },
+  "scripts": {
+    "dev": "cross-env env=DEV node ./script/main.js",
+    "build": "cross-env env=BUILD node ./script/main.js"
+  },
+  "devDependencies": { "cross-env": "^7.0.3" }
 }
